FlarmId: change "long" to "uint32_t"
The FLARM traffic ID is parsed from a 6 digit hex number.  Don't waste
memory on a "long" which is 64 bit wide on some platforms; hard-code
it to use 32 bits.

+++ src/FLARM/FlarmId.cpp
@@ -50,6 +50,6 @@ FlarmId::parse(const TCHAR *input, TCHAR **endptr_r)
 const TCHAR *
 FlarmId::format(TCHAR *buffer) const
 {
-  _stprintf(buffer, _T("%lX"), value);
+  _stprintf(buffer, _T("%lX"), (unsigned long)value);
   return buffer;
 }
